+++ src/hhu/propra2013/gruppe55/GameEventListener.java
@@ -3,10 +3,10 @@ package hhu.propra2013.gruppe55;
 public interface GameEventListener {
 	
 	// Vorlauefiges Beute-Event
-	public void newTreasure(int x, int y);
+	public void newTreasure(double x, double y);
 	
 	// Boss droppt Ziel
-	public void newGoal(int x, int y);
+	public void newGoal(double x, double y);
 	
 	// Geschosse werden geschossen
 	public void shootProjectile(Projectile p);
