+++ src/hhu/propra2013/gruppe55/GoalObject.java
@@ -2,7 +2,7 @@ package hhu.propra2013.gruppe55;
 
 public class GoalObject extends DungeonObject {
 		
-	public GoalObject(int x, int y) {
+	public GoalObject(double x, double y) {
 		super(x,y);
 		state[0].changeImg(Data_Img.goal);			  // Aussehen des Ziels
 		//state[0].massive=false;				  // Ziel wird begehbar
